change delete meal controller to accept array of meal ids

+++ models/mealModel.js
@@ -24,15 +24,15 @@ Meal.init(
       allowNull: false,
     },
     carbs: {
-      type: DataTypes.FLOAT,
+      type: DataTypes.INTEGER,
       allowNull: false,
     },
     protein: {
-      type: DataTypes.FLOAT,
+      type: DataTypes.INTEGER,
       allowNull: false,
     },
     fats: {
-      type: DataTypes.FLOAT,
+      type: DataTypes.INTEGER,
       allowNull: false,
     },
   },
